refactor: move superset-ui/chart-composition to core
