--- packages/superset-ui-chart-composition/README.md
@@ -1,28 +1,0 @@
-## @superset-ui/chart-composition
-
-[![Version](https://img.shields.io/npm/v/@superset-ui/chart-composition.svg?style=flat)](https://img.shields.io/npm/v/@superset-ui/chart-composition.svg?style=flat)
-[![David (path)](https://img.shields.io/david/apache-superset/superset-ui.svg?path=packages%2Fsuperset-ui-chart-composition&style=flat-square)](https://david-dm.org/apache-superset/superset-ui?path=packages/superset-ui-chart-composition)
-
-Description
-
-#### Example usage
-
-```js
-import {
-  ChartFrame,
-  TooltipFrame,
-  TooltipTable,
-  WithLegend,
-} from '@superset-ui/chart-composition';
-```
-
-#### API
-
-`fn(args)`
-
-- Do something
-
-### Development
-
-`@data-ui/build-config` is used to manage the build configuration for this package including babel
-builds, jest testing, eslint, and prettier.
