--- packages/superset-ui-chart-composition/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@superset-ui/chart-composition",
-  "version": "0.14.11",
-  "description": "Superset UI chart-composition",
-  "sideEffects": false,
-  "main": "lib/index.js",
-  "module": "esm/index.js",
-  "files": [
-    "esm",
-    "lib"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/apache-superset/superset-ui.git"
-  },
-  "keywords": [
-    "superset"
-  ],
-  "author": "Superset",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/apache-superset/superset-ui/issues"
-  },
-  "homepage": "https://github.com/apache-superset/superset-ui#readme",
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@types/react": "^16.9.43",
-    "@vx/responsive": "^0.0.197",
-    "@superset-ui/core": "^0.15.0",
-    "csstype": "^2.6.4"
-  },
-  "peerDependencies": {
-    "react": "^16.13.1"
-  }
-}
