--- packages/superset-ui-chart-composition/test/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "composite": false,
-    "emitDeclarationOnly": false,
-    "noEmit": true,
-    "rootDir": "."
-  },
-  "extends": "../../../tsconfig.options.json",
-  "include": [
-    "**/*",
-    "../types/**/*",
-    "../../../types/**/*"
-  ],
-  "references": [
-    {
-      "path": ".."
-    }
-  ]
-}
