--- packages/superset-ui-chart-composition/types/@vx/responsive/index.d.ts
@@ -1,11 +1,0 @@
-declare module '@vx/responsive' {
-  import React from 'react';
-
-  interface ParentSizeProps {
-    debounceTime?: number;
-    children: (renderProps: { width: number; height: number }) => React.ReactNode;
-  }
-
-  // eslint-disable-next-line import/prefer-default-export
-  export const ParentSize: React.ComponentType<ParentSizeProps>;
-}
