+++ packages/superset-ui-core/package.json
@@ -40,6 +40,9 @@
     "@types/d3-time-format": "^2.1.0",
     "@types/d3-interpolate": "^1.3.1",
     "@types/d3-scale": "^2.1.1",
+    "@types/react": "*",
+    "@vx/responsive": "^0.0.197",
+    "csstype": "^2.6.4",
     "d3-interpolate": "^1.4.0",
     "d3-scale": "^3.0.0",
     "d3-format": "^1.3.2",
@@ -51,5 +54,8 @@
     "fetch-retry": "^4.0.1",
     "pretty-ms": "^7.0.0",
     "whatwg-fetch": "^3.0.0"
+  },
+  "peerDependencies": {
+    "react": "^16.13.1"
   }
 }
