+++ packages/superset-ui-core/types/resize-observer-polyfill.d.ts
@@ -1,0 +1,4 @@
+declare module 'resize-observer-polyfill' {
+  // eslint-disable-next-line import/prefer-default-export
+  export function triggerResizeObserver(entries?: ResizeObserverEntry[]): void;
+}
