+++ plugins/legacy-plugin-chart-time-table/package.json
@@ -33,7 +33,7 @@
     "@superset-ui/chart-controls": "^0.14.3",
     "@superset-ui/core": "^0.15.0",
     "@types/d3-scale": "^2.0.2",
-    "@types/react": "^16.9.43",
+    "@types/react": "*",
     "d3-scale": "^3.2.1",
     "emotion-theming": "^10.0.27",
     "moment": "^2.26.0",
