+++ plugins/plugin-chart-table/package.json
@@ -32,7 +32,7 @@
     "@superset-ui/core": "^0.15.0",
     "@types/d3-array": "^2.0.0",
     "@types/match-sorter": "^4.0.0",
-    "@types/react": "^16.9.43",
+    "@types/react": "*",
     "@types/react-table": "^7.0.19",
     "d3-array": "^2.4.0",
     "match-sorter": "^4.1.0",
