+++ plugins/preset-chart-xy/package.json
@@ -30,7 +30,6 @@
   },
   "dependencies": {
     "@superset-ui/chart": "^0.14.0",
-    "@superset-ui/chart-composition": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
     "@data-ui/theme": "^0.0.84",
