+++ temporary-plugins/package.json
@@ -52,7 +52,7 @@
     "@types/enzyme": "^3.10.3",
     "@types/jest": "^25.1.3",
     "@types/jsdom": "^12.2.4",
-    "@types/react": "^16.9.43",
+    "@types/react": "*",
     "@types/react-test-renderer": "^16.9.2",
     "csstype": "^2.6.3",
     "enzyme": "^3.11.0",
