+++ temporary-plugins/superset-ui-plugins-demo/package.json
@@ -37,7 +37,7 @@
     "@storybook/addon-links": "^5.0.9",
     "@storybook/addons": "^5.0.9",
     "@storybook/react": "^5.0.9",
-    "@types/react": "^16.9.43",
+    "@types/react": "^16.9.48",
     "@types/storybook__react": "4.0.2",
     "bootstrap": "^3.4.1",
     "jquery": "^3.4.1",
@@ -48,7 +48,6 @@
     "@babel/core": "^7.4.3",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@superset-ui/chart": "^0.14.0",
-    "@superset-ui/chart-composition": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
     "babel-loader": "^8.0.5",
     "gh-pages": "^2.0.1",
